+++ src/sbitx_sound.c
@@ -1081,7 +1081,7 @@ void *sound_thread_function(void *ptr){
 //  printf("opening loopback on plughw:CARD=Loopback,DEV=0 sound card\n");
 
 	for (i = 0; i < 10; i++){
-		if(sound_start_loopback_play("plughw:CARD=Loopback,DEV=0") == 0)
+		if(sound_start_loopback_play("plughw:CARD=1,DEV=0") == 0)
 			break;
 		fprintf(stderr, "*Error opening Loopback Play device");
 		delay(1000);
@@ -1108,7 +1108,7 @@ void *loopback_thread_function(void *ptr){
 
 	int i = 0;
 	for (i = 0; i < 10; i++){
-		if (sound_start_loopback_capture("plughw:CARD=Loopback_1,DEV=1") == 0)
+		if (sound_start_loopback_capture("plughw:CARD=2,DEV=1") == 0)
 			break;
 		fprintf(stderr, "*Error opening Loopback Capture device");
 		delay(1000);